requirements.txt
# -------------------------
# Core backend
# -------------------------
fastapi
uvicorn[standard]

# -------------------------
# LangChain / LangGraph
# -------------------------
langchain
langgraph

# -------------------------
# File ingestion & OCR
# -------------------------
pdfplumber
pytesseract
pillow
python-docx

# -------------------------
# Text processing
# -------------------------
regex

# -------------------------
# Databases
# -------------------------
firebase-admin
neo4j
qdrant-client

# -------------------------
# Utilities
# -------------------------
python-dotenv
pydantic
requests

opencv-python<4.9
numpy<2

opencv-python
scispacy
https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_ner_bc5cdr_md-0.5.4.tar.gz
pdfplumber
pytesseract